perf(generate): shrink `Symbol::index` from `usize` to `u32`
This cuts `Symbol`'s size from 16 to 8 bytes. Results in a ~2-7% wall time
reduction, and ~7-10% peak RSS reduction.

+++ crates/generate/src/build_tables/token_conflicts.rs
@@ -318,7 +318,7 @@ fn get_following_chars(
             let mut chars = CharacterSet::empty();
             for token in following_tokens.iter() {
                 if token.is_terminal() {
-                    chars = chars.add(&starting_chars[token.index]);
+                    chars = chars.add(&starting_chars[token.index as usize]);
                 }
             }
             chars
